Switch from `treefmt` -> `pre-commit` (#144)
I feel that the quality of `treemt` has lately declined. For eg., recent
versions started dumping a bunch of needless noise (why is it now trying
to format all files in repo?):


![image](https://github.com/user-attachments/assets/9aa15526-1081-41f1-85f4-7dc2034b24d2)

which then required more configuration in Nix to fix:


https://github.com/srid/haskell-template/commit/d8c7e63731d9b63ef5937cb3fdc8694046ea6c39

And this unnecessarily extra config will vary across projects, depending
on the files they have.

---

We strictly don't need treefmt though (besides, [not
everyone](https://x.com/sridca/status/1808605343674450157) is allowed to
contribute) because that's what `pre-commit run -a` is for.

To that end, this PR gets rid of `treefmt` and switches to
https://github.com/cachix/git-hooks.nix (the Nix module for
https://pre-commit.com/).

Functionality more or less remains the same - except for losing `just`
formatter (which can be added to git-hooks.nix) and of course gaining
git pre-commit hooks setup in devShell.